Login Page Work  - March 2017
The login form in java/html is added as of line 4 in the homepage.html document. This login allows for a user to put in a set username and password and the page will confirm or deny whether the username and password ae correct which allows the user to sign in on the page.