Update Login Work to Submit - March 2017
Please look at lines 1-32 (everything in the head) for my rudimentary login form. To initiate login prompts, click the sign in button. The prompts will notify you if the username/passwords are correct. After that, follow the instructions and click the second button to finish the sign in, the username will then display above step 2 text.